+++ todosync/__main__.py
@@ -6,10 +6,9 @@ from todosync import todosync
 
 @click.command()
 @click.option('--dry-run', is_flag=True, help='Do not commit the changes.')
-@click.option('--config', help='Path to the configuration file.')
-def execute(dry_run: bool, config: str):
+def execute(dry_run: bool):
     """Synchronize your Git{hub,lab} issues with Todoist."""
-    todosync.synchronize(dry_run, config)
+    todosync.synchronize(dry_run)
 
 
 if __name__ == '__main__':